real_time_pipeline/processor/requirements.txt
functions-framework==3.4.0
google-cloud-bigquery==2.34.3
google-cloud-pubsub==2.13.0
google-cloud-storage==2.7.0